% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elsa.else.R
\name{Elsa.else}
\alias{Elsa.else}
\title{Elsa.else Function}
\usage{
Elsa.else(df_you, id_name, group_name, fun_name, fun_type = 1, prob = 0.25)
}
\arguments{
\item{df_you}{数据集}

\item{id_name}{须生成id的列}

\item{group_name}{是分组的变量,若需要group的有多个组,如city-year组，只需给新的组合city-year打上id即可}

\item{fun_name}{待处理的列}

\item{fun_type}{设定了6种，1.sum，3.mean，3.median，4.sd，5.quantile，并可使用prob参数；6.prod}
}
\description{
用于计算分组后除自身外的其他观测值的汇总指标：例如求和、平均值、分位数等
}
\examples{
Elsa(df_you = df2,
id_name =行业,
group_name = newid,
fun_name =行业j就业人数,
fun_type = 1)

}
\author{
Yu Li_XMU
}
\keyword{df}
