\docType{class}
\name{Entrogram-class}
\alias{Entrogram-class}
\alias{Variogram-class}
\alias{Correlogram-class}
\alias{data.frameORmatrix-class}
\alias{matrixORnull-class}

\title{An S4 class for Entrogram}
\description{
An S4 class representing Entrogram dataset generated by \code{entrogram} function
}
\section{Slots}{

\describe{
\item{\code{width}}{the width of susequent distance intervals}

\item{\code{cutoff}}{spatial separation distance up to which point pairs are included in entrogram estimates}

\item{\code{entrogramCloud}}{A matrix containing the elsa for each location and the neighbourhood distance}

\item{\code{entrogram}}{A data.frame containing mean elsa values at each distance interval}
}}

\references{

Naimi, B., Hamm, N. A., Groen, T. A., Skidmore, A. K., Toxopeus, A. G., & Alibakhshi, S. (2019). ELSA: Entropy-based local indicator of spatial association. Spatial statistics, 29, 66-88.
}

\author{Babak Naimi \email{naimi.b@gmail.com}

\url{http://r-gis.net}
}

\keyword{spatial}