\docType{class}
\name{neighbours-class}
\alias{neighbours-class}

\title{An S4 class of neighbour features}
\description{
An S4 class representing a list of neighbour features generated by \code{dneigh} function
}
\section{Slots}{

\describe{
\item{\code{distance1}}{the distance from which the neighbour features are seeked}

\item{\code{distance2}}{the distance up to which the neighbour features are seeked}

\item{\code{neighbours}}{A list containing numeric vectors each specifies which the neighbours for the corresponsing feature}

}}

\references{

Naimi, B., Hamm, N. A., Groen, T. A., Skidmore, A. K., Toxopeus, A. G., & Alibakhshi, S. (2019). ELSA: Entropy-based local indicator of spatial association. Spatial statistics, 29, 66-88.
}

\author{Babak Naimi \email{naimi.b@gmail.com}

\url{http://r-gis.net}
}

\keyword{spatial}