\name{elsa.test}

\alias{elsa.test}
\alias{elsa.test,RasterLayer-method}
\alias{elsa.test,Spatial-method}



\title{Elsa test for local spatial autocorrelation}

\description{
This function can be used to convert a data.frame including the values specifying the differences (digree of dissimilarities) between the classes (categories) in a categoricl map, to a list.
}

\usage{
elsa.test(x, d, n, method, null, nc, categorical, dif,...)
}

\arguments{
  \item{x}{ A Raster or Spatial* dataset}
  \item{d}{the local distance, or an object of class neighbours created by dneigh function}
  \item{n}{number of simulation, default is 999 for small datasets, and 99 for large datasets}
  \item{null}{ Optional, a null distribution of data (a Raster if x is Raster or a numerical vector if x is either Raster or Spatial dataset ); if not provided, a null distribution is generated by the function}
  \item{method}{resampling method for nonparametric simulation, can be either 'boot' (bootstraping; default) or 'perm' (permutation) }
  \item{nc}{number of classes (only if x is a continuous variable); if not specified, it is estimated using nclass function}
  \item{categorical}{logical, specifies whether x is a categorical; if not specified, it is guessed by the function}
  \item{dif}{the level of dissimilarities between different categories (only if x is a categorical variable); see \code{\link[elsa]{dif2list}} for more details}
  \item{...}{Aditional arguments passed to writeRaster function (applied only when x is Raster)}
}
\details{
This function test how significant the local spatial autocorrelation is at each location, so it generates a p-value at each location through a Monte Carlo simulation and a non-parametric approach. See the reference for the details about the method.

If \code{null} distribution is not provided, the function generates a null distribution by drawing a sample at each location from a unique set of m values in dataset (categories) each having a probability of \eqn{1/m}. 

}

\value{
An object same as the input (\code{x})

}

\references{
Naimi. B. et al. (under review) ELSA: An Entropy-based Local indicators of Spatial Association, Geographical Analysis;

}

\author{Babak Naimi \email{naimi.b@gmail.com}

\url{http://r-gis.net}
}


\examples{
\dontrun{
file <- system.file('external/dem_example.grd',package='elsa')

r <- raster(file)

plot(r,main='a continuous raster map')

et <- elsa.test(r,d=2000,n=99)

plot(et)

}

}

\keyword{spatial}
