\docType{class}
\name{neighbours-class}
\alias{neighbours-class}

\title{An S4 class of neighbour features}
\description{
An S4 class representing a list of neighbour features generated by \code{dneigh} function
}
\section{Slots}{

\describe{
\item{\code{distance1}}{the distance from which the neighbour features are seeked}

\item{\code{distance2}}{the distance up to which the neighbour features are seeked}

\item{\code{neighbours}}{A list containing numeric vectors each specifies which the neighbours for the corresponsing feature}

}}

\references{
Naimi, B., Araujo, M.B. (2016) sdm: a reproducible and extensible R platform for species distribution modelling, Ecography, DOI: 10.1111/ecog.01881
}

\author{Babak Naimi \email{naimi.b@gmail.com}

\url{http://r-gis.net}
}

\keyword{spatial}